requirement.txt
cv2
imutils
matplotlib
scikit-image
scikit-learn
Flask
mysql-connector-python